fix types

+++ types/types.go
@@ -78,19 +78,19 @@ func ParquetTypeToGoReflectType(pT *parquet.Type, cT *parquet.ConvertedType, rT 
 			return reflect.TypeOf(int32(0))
 
 		} else if *pT == parquet.Type_INT32 && *cT == parquet.ConvertedType_INT_8 {
-			return reflect.TypeOf(int8(0))
+			return reflect.TypeOf(int32(0))
 
 		}else if *pT == parquet.Type_INT32 && *cT == parquet.ConvertedType_INT_16 {
-			return reflect.TypeOf(int16(0))
+			return reflect.TypeOf(int32(0))
 
 		}else if *pT == parquet.Type_INT32 && *cT == parquet.ConvertedType_INT_32 {
 			return reflect.TypeOf(int32(0))
 
 		}else if *pT == parquet.Type_INT32 && *cT == parquet.ConvertedType_UINT_8 {
-			return reflect.TypeOf(uint8(0))
+			return reflect.TypeOf(uint32(0))
 
 		}else if *pT == parquet.Type_INT32 && *cT == parquet.ConvertedType_UINT_16 {
-			return reflect.TypeOf(uint16(0))
+			return reflect.TypeOf(uint32(0))
 
 		}else if *pT == parquet.Type_INT32 && *cT == parquet.ConvertedType_UINT_32 {
 			return reflect.TypeOf(uint32(0))
@@ -139,11 +139,11 @@ func ParquetTypeToGoReflectType(pT *parquet.Type, cT *parquet.ConvertedType, rT 
 			return reflect.TypeOf(&v)
 
 		} else if *pT == parquet.Type_INT32 && *cT == parquet.ConvertedType_INT_8 {
-			v := int8(0)
+			v := int32(0)
 			return reflect.TypeOf(&v)
 
 		} else if *pT == parquet.Type_INT32 && *cT == parquet.ConvertedType_INT_16 {
-			v := int16(0)
+			v := int32(0)
 			return reflect.TypeOf(&v)
 
 		} else if *pT == parquet.Type_INT32 && *cT == parquet.ConvertedType_INT_32 {
@@ -151,11 +151,11 @@ func ParquetTypeToGoReflectType(pT *parquet.Type, cT *parquet.ConvertedType, rT 
 			return reflect.TypeOf(&v)
 		
 		} else if *pT == parquet.Type_INT32 && *cT == parquet.ConvertedType_UINT_8 {
-			v := uint8(0)
+			v := uint32(0)
 			return reflect.TypeOf(&v)
 
 		} else if *pT == parquet.Type_INT32 && *cT == parquet.ConvertedType_UINT_16 {
-			v := uint16(0)
+			v := uint32(0)
 			return reflect.TypeOf(&v)
 
 		} else if *pT == parquet.Type_INT32 && *cT == parquet.ConvertedType_UINT_32 {
